Add frame-lost strategy, remove dead config, fix kbps overflow
Frame-lost strategy (SetFrameLostStrategy):
- Safety net: drops frames when encoder instantaneous bitrate exceeds
  120% of configured target, preventing pipeline stalls on overflow
- NORMAL mode (PSKIP not supported on Infinity6E kernel)
- Gated by video0.frameLost config (default: true), with camelCase
  API alias. Dual VENC (recording) always enabled regardless of config
- Called after StartRecvPic (kernel requires channel in receive state)
- Struct layout verified against SDK disassembly (4 words, 16 bytes)
- Device-verified: 89fps@90fps and 119fps@120fps maintained, zero
  frame drops under normal operation

Remove dead config fields:
- slicesEnabled/sliceSize: firmware accepts API but never produces
  multiple slice NALUs (verified via packetizer NALU counts)
- lowDelay: VIF frameLineCnt alone gives +108us (noise). Full REALTIME
  VPE→VENC pipeline needed for real gain — documented in
  documentation/LOW_DELAY_PIPELINE.md as high-priority future work
- Cleaned across 11 files: struct, defaults, parse, serialize, API
  fields, alias, tests, scripts, default config, docs

Fix kbps verbose overflow on 32-bit ARM:
- interval_bytes * 8000UL overflows uint32 above ~4.3 Mbps
- At 13 Mbps actual, displayed 392 kbps (confirmed by math: overflow
  wraps to exactly the observed value)
- Fix: cast to unsigned long long for 64-bit arithmetic
- Pre-existing bug since initial commit, not a regression

Scene-change IDR investigated and removed:
- startQual field in i6_venc_strminfo is always 0 on Infinity6E —
  the VENC driver does not populate per-frame QP data
- Full implementation with QP-based I-frame detection and 2-frame
  cooldown was developed and reviewed but removed as dead code
- Remains a future opportunity if per-frame QP becomes available

+++ src/stream_metrics.c
@@ -47,8 +47,8 @@ int stream_metrics_sample(StreamMetricsState *state,
 	sample->elapsed_ms = elapsed_ms;
 	sample->uptime_s = now->tv_sec - state->ts_start.tv_sec;
 	sample->fps = (unsigned int)((state->interval_frames * 1000L) / elapsed_ms);
-	sample->kbps = (unsigned int)((state->interval_bytes * 8000UL) /
-		((unsigned long)elapsed_ms * 1024UL));
+	sample->kbps = (unsigned int)(((unsigned long long)state->interval_bytes * 8000ULL) /
+		((unsigned long long)elapsed_ms * 1024ULL));
 	if (state->interval_frames > 0) {
 		sample->avg_bytes = (unsigned int)(state->interval_bytes /
 			state->interval_frames);
